working docker build .

+++ creator-node/scripts/up.sh
@@ -11,12 +11,14 @@ mkdir -p compose/env/tmp/file-storage-${1}
 . compose/env/tmp/shellEnv${1}.sh
 
 # build docker image without node_modules
-mv node_modules /tmp/cn-node_modules
-time docker build --progress=plain .
-time docker-compose -f compose/docker-compose.yml build
-[ -d node_modules ] && mv node_modules/* /tmp/cn-node_modules/ || true
-rm -rf node_modules
-mv /tmp/cn-node_modules node_modules
+if [[ "${1}" == 1 ]]; then
+    mv node_modules /tmp/cn-node_modules
+    time docker build --progress=plain .
+    time docker-compose -f compose/docker-compose.yml build
+    [ -d node_modules ] && mv node_modules/* /tmp/cn-node_modules/ || true
+    rm -rf node_modules
+    mv /tmp/cn-node_modules node_modules
+fi
 
 mkdir -p compose/env/tmp/file-storage-${1}
 . compose/env/tmp/shellEnv${1}.sh
